requirements.txt
Flask==2.2.3
keras==2.11.0
numpy==1.21.4
Pillow==9.5.0
tensorflow-cpu==2.11.0
